allowed weight to be float

+++ api_packing/v1/models/variants_request_model.py
@@ -11,7 +11,7 @@ class ArticleModel(BaseModel):
     width: int = Field(description="Article width", gt=0)
     length: int = Field(description="Article length", gt=0)
     height: int = Field(description="Article height", gt=0)
-    weight: Optional[int] = Field(description="Article weight", ge=0, default=0)
+    weight: Optional[float] = Field(description="Article weight", ge=0, default=0)
     amount: int = Field(description="Amount of articles", gt=0)
 
 
@@ -22,7 +22,7 @@ class ColliDetailsModel(BaseModel):
     length: int = Field(description="Length of the bins to be packed", gt=0)
     height: int = Field(description="Height of the bins to be packed", gt=0)
     max_collis: int = Field(description="Maximum number of bins to be packed", gt=0)
-    max_weight: Optional[int] = Field(
+    max_weight: Optional[float] = Field(
         description="Maximum weight of the bins to be packed", ge=0, default=None
     )
